deprecate: input.prompt_buffer option
Even once the fixes to the prompt buffer come out in the next version of
Neovim, I'm really not sure what advantages we get from using the prompt
buffer. And we will continue to have disadvantages because it does *not*
function like a normal buffer. If anyone is using it and has a good
reason for it, please let me know. Otherwise, I'll be removing that
functionality in a later commit so I don't have to support two
implementations.