fix!: drop specialized text highlight groups (#30)
Removing DressingInputText and DressingSelectText highlight groups that
were added for #8. I added them before I learned about winhighlight, and
it turns out that winhighlight is the better way to do this. Previously
I was using `nvim_buf_add_highlight` to directly add the highlight
group, but that has multiple downsides (can't highlight bg of regions
with no text, overrides other relevate highlight groups e.g.
CursorLine). None of that is needed because Neovim will be setting the
highlight group of this text to NormalFloat, and if the user wants to
customize it we already expose an easy option in config to do
`winhighlight = "NormalFloat:MyCustomHighlightGroup"`.