Do not display item number for fzf_lua
I struggled to disable the item numbers until I found out it was from dressing and not fzf_lua that defines behavior. It should be consistent with the other backends which do not display the item number. If the user wishes so, it is easy to do in the select function.