refactor!: deprecate the "anchor" config option (#106)
This was left over from before we had the "override" function. Now, it
doesn't make much sense to set the anchor as an option because so much
of the rest of the layout logic depends on the anchor being the default.
If a user wants to customize the layout, the override function is the
way to do it.